Remove duplicated extension helper Signed-off-by: Zachary Nixon <nixozach@amazon.com>
